pyproject.toml
[project]
name = "frag-console"
version = "0.1.0"
description = "Console for FRAG Dataspace."
authors = [
    { name = "mitk23" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "pydantic>=2.10.1",
    "python-keycloak>=4.7.2",
    "pyyaml>=6.0.2",
    "qdrant-client>=1.12.1",
    "tqdm>=4.67.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/default", "src"]

[tool.ruff]
line-length = 120
indent-width = 4
extend-include = ["*.ipynb"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
